% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansClusterPlot.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{Perform K-means Clustering}
\usage{
kmeans_clustering(data, k)
}
\arguments{
\item{data}{A data frame containing numeric variables.}

\item{k}{The number of clusters.}
}
\value{
A list object of class `kmeans` (the result of the `kmeans()` function).
}
\description{
Perform K-means Clustering
}
