% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansClusterPlot.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot K-means Clustering Results}
\usage{
plot_clusters(data, kmeans_result)
}
\arguments{
\item{data}{A data frame containing numeric variables.}

\item{kmeans_result}{The result of the `kmeans_clustering()` function (a kmeans object).}
}
\value{
A scatter plot with clusters.
}
\description{
Plot K-means Clustering Results
}
